refactor region parser, remove 'unknown' type, move calculator specs around
