+++ Specs/Controllers/Stream/Calculators/CategoryHeaderCellSizeCalculatorSpec.swift
