+++ Specs/Controllers/Stream/Calculators/StreamNotificationCellSizeCalculatorSpec.swift
@@ -8,20 +8,6 @@ import Nimble
 
 
 class StreamNotificationCellSizeCalculatorSpec: QuickSpec {
-    class MockUIWebView: UIWebView {
-        var mockHeight: CGFloat = 50
-
-        override func loadHTMLString(_ html: String, baseURL: URL?) {
-            delegate?.webViewDidFinishLoad?(self)
-        }
-
-        override func stringByEvaluatingJavaScript(from js: String) -> String? {
-            if js.contains("offsetWidth") { return "\(frame.size.width)" }
-            if js.contains("offsetHeight") { return "\(mockHeight)" }
-            return super.stringByEvaluatingJavaScript(from: js)
-        }
-    }
-
     override func spec() {
         describe("StreamNotificationCellSizeCalculator") {
             let user: User = stub([:])
